+++ pkg/distributor/distributor.go
@@ -786,6 +786,10 @@ func (d *Distributor) Push(ctx context.Context, req *cortexpb.WriteRequest) (*co
 		if sp := opentracing.SpanFromContext(ctx); sp != nil {
 			localCtx = opentracing.ContextWithSpan(localCtx, sp)
 		}
+		// Get any HTTP headers that are supposed to be added to logs and add to localCtx for later use
+		if headerMap := util_log.HeaderMapFromContext(ctx); headerMap != nil {
+			localCtx = util_log.ContextWithHeaderMap(localCtx, headerMap)
+		}
 
 		// Get clientIP(s) from Context and add it to localCtx
 		localCtx = util.AddSourceIPsToOutgoingContext(localCtx, source)